api/requirements.txt
azure-functions
azure-cosmos
python-dotenv
requests
pandas
numpy
torch
scikit-learn
fastapi
pydantic
uvicorn